DebugをCommonにおいてプロジェクト側はRootを継承して使用するように変更
