+++ Assets/App/Scripts/_Debug/DebugRootMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a1cb326989e74e679acb3e82a46617f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
